Remove go1.5-go1.6 vendor bridge - and duplicate go/build functions
The original purpose of this package was to provide an implementation
of the go/build package that would bridge the transition of vendoring
between go1.5 and go1.6.  This was done mostly to support my fork of
GoSublime.

That transition is now long over and this package is now mostly used
to quickly determine if files should be imported and what build tags
they have, if any.

This commit removes the legacy code - greatly reducing the size of
the code base.